MueLu: fix oneapi compilation error
Move RepartitionUtilities class to its own file for the ETI system.
