+++ packages/muelu/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -91,6 +91,7 @@ APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ReorderBlockAFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RepartitionFactory-.?ifdef.HAVE_MPI )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RepartitionBlockDiagonalFactory-.?ifdef.HAVE_MPI )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RepartitionHeuristicFactory-.?ifdef.HAVE_MPI )
+APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RepartitionUtilities-.?ifdef.HAVE_MPI )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ReplicatePFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RigidBodyModeFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SaPFactory )
