+++ packages/muelu/src/Utils/ForwardDeclaration/MueLu_RepartitionUtilities_fwd.hpp
@@ -1,0 +1,27 @@
+// @HEADER
+// *****************************************************************************
+//        MueLu: A package for multigrid based preconditioning
+//
+// Copyright 2012 NTESS and the MueLu contributors.
+// SPDX-License-Identifier: BSD-3-Clause
+// *****************************************************************************
+// @HEADER
+
+#ifndef MUELU_REPARTITIONUTILITIES_FWD_HPP
+#define MUELU_REPARTITIONUTILITIES_FWD_HPP
+
+#include "MueLu_ConfigDefs.hpp"
+#ifdef HAVE_MPI
+
+namespace MueLu {
+template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+class RepartitionUtilities;
+}
+
+#ifndef MUELU_REPARTITIONUTILITIES_SHORT
+#define MUELU_REPARTITIONUTILITIES_SHORT
+#endif
+
+#endif
+
+#endif  // MUELU_REPARTITIONUTILITIES_FWD_HPP
